Factor out code from Test_RegisterElements
- The factored out code will be common to other observer tests
- Adds a missing include to Actions.hpp
